etl/requirements.txt
psycopg2-binary
python-dotenv
numpy
pandas
pypdf
